+++ CHANGELOG.md
@@ -7,3 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [0.1.0] - 2020-05-16
 ### Added
  - Packing tool that supports GPU and CPU channel packing.
+
+## [0.2.0] - 2024-02-01
+
+### Added
+ - Ability to choose which channel each source will pull from
+
+### Fixed
+ - Correct srgb settings now used when creating textures